Fixed and added aadhaar checking API.
