+++ app/build.gradle.kts
@@ -1,14 +1,15 @@
 plugins {
     id("com.android.application")
     id("org.jetbrains.kotlin.android")
+    id("com.google.gms.google-services")
 }
 
 android {
-    namespace = "com.example.ocacapp"
+    namespace = "com.jetpack.ocac"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.example.ocacapp"
+        applicationId = "com.jetpack.ocac"
         minSdk = 24
         targetSdk = 34
         versionCode = 1
@@ -64,6 +65,7 @@ dependencies {
     implementation(platform("androidx.compose:compose-bom:2024.06.00"))
     implementation(platform("androidx.compose:compose-bom:2024.06.00"))
     implementation(platform("androidx.compose:compose-bom:2023.08.00"))
+    implementation("com.google.firebase:firebase-messaging-ktx:24.0.0")
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.2.1")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.6.1")
@@ -84,4 +86,6 @@ dependencies {
     implementation("androidx.multidex:multidex:2.0.1")
 /// Loading network image from URL
     implementation ("io.coil-kt:coil-compose:2.1.0")
+    implementation("com.google.firebase:firebase-analytics")
+    implementation(platform("com.google.firebase:firebase-bom:33.1.2"))
 }
