--- app/src/main/java/com/example/ocacapp/GridMenu.kt
@@ -1,6 +1,0 @@
-package com.example.ocacapp
-
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.painter.Painter
-
-data class GridMenu(val modifier: Modifier, val icon: Painter, val title: String)
