--- app/src/main/java/com/example/ocacapp/MenuItem.kt
@@ -1,6 +1,0 @@
-package com.example.ocacapp
-
-import androidx.compose.ui.graphics.painter.Painter
-
-
-data class MenuItem(val id: String, val title: String, val icon: Painter)
