--- app/src/main/java/com/example/ocacapp/Model/LanguageModel.kt
@@ -1,14 +1,0 @@
-package com.example.ocacapp.Model
-
-data class LanguageModel(
-    val `data`: List<LanguageData>,
-    val error: Int,
-    val lang: String,
-    val message: String,
-    val success: Int
-)
-
-data class LanguageData(
-    val name: String,
-    val value: String
-)
