+++ app/src/main/java/com/jetpack/ocacapp/GridMenu.kt
@@ -1,0 +1,6 @@
+package com.jetpack.ocac
+
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.painter.Painter
+
+data class GridMenu(val modifier: Modifier, val icon: Painter, val title: String)
