+++ app/src/main/java/com/jetpack/ocacapp/MenuItem.kt
@@ -1,0 +1,6 @@
+package com.jetpack.ocac
+
+import androidx.compose.ui.graphics.painter.Painter
+
+
+data class MenuItem(val id: String, val title: String, val icon: Painter)
